autoont/pyproject.toml
[project]
name = "autoont"
version = "0.1.0"
description = "Automatic ontology generation from images via VLM observations"
requires-python = ">=3.11"
dependencies = [
    "ollama>=0.4.0",
    "pillow>=10.0.0",
    "kg-gen @ file:///Users/rob/repos/stair-lab/kg-gen",
]

[tool.uv]
dev-dependencies = []
